apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.4
pydantic-settings==2.6.1
httpx==0.27.2
groq==0.4.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
requests==2.32.3
firebase-admin==6.6.0
ollama==0.4.5
websockets==14.1
python-dotenv==1.0.1
aiofiles==24.1.0
pillow==11.0.0
PyPDF2==3.0.1